requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.1
pytesseract==0.3.10
Pillow==10.3.0
pydantic==2.7.1
